lsj/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1.13
django-cors-headers==3.1.0
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
idna==2.8
jedi==0.15.1
mixins==0.1.1
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.8
requests==2.22.0
urllib3==1.25.6
